Extract run marker overflow preflight
